cli: Fix console crash when using type function
The console-only type function allows interrogation of any value's type.
An implementation quirk is that we use a cty.Mark to allow the console
to display this type information without the usual HCL quoting. For
example:

> type("boop")
string

instead of:

> type("boop")
"string"

Because these marks can propagate when used in complex expressions,
using the type function as part of constructing a non-string value could
result in this "print as raw" mark being attached to a collection. When
this happened, it would result in a crash when we tried to iterate over
a marked value.

This commit fixes this by stripping the raw mark from any non-string
types before attempting to format the output. We are careful not to
strip other marks, to ensure we don't accidentally reveal sensitive
information by doing so.